+++ CHANGELOG.md
@@ -16,10 +16,14 @@
 - Added `spicyweb\embeddedassets\Service::EVENT_BEFORE_REQUEST`
 
 ### Changed
+- `spicyweb\embeddedassets\models\EmbeddedAsset::$imageHeight` will not be set on newly created or refreshed embedded assets
+- `spicyweb\embeddedassets\models\EmbeddedAsset::$imageWidth` will not be set on newly created or refreshed embedded assets
 - `spicyweb\embeddedassets\models\EmbeddedAsset::$images` now returns an array of image URLs, rather than an array of key/value pairs representing image data, and will not be set on newly created or refreshed embedded assets
 - `spicyweb\embeddedassets\models\EmbeddedAsset::$tags` will not be set on newly created or refreshed embedded assets
 
 ### Deprecated
+- Deprecated `spicyweb\embeddedassets\models\EmbeddedAsset::$imageHeight`, due to being removed in Embed 4
+- Deprecated `spicyweb\embeddedassets\models\EmbeddedAsset::$imageWidth`, due to being removed in Embed 4
 - Deprecated `spicyweb\embeddedassets\models\EmbeddedAsset::$images`; use `$image` instead
 - Deprecated `spicyweb\embeddedassets\models\EmbeddedAsset::$tags`; use `$keywords` instead